Move awa from functions to classes
