+++ cdk/lib/bitbucket-metrics.ts
@@ -2,7 +2,6 @@ import {Construct} from 'constructs';
 import {HttpApi} from 'aws-cdk-lib/aws-apigatewayv2';
 import {Stack, Stage} from 'aws-cdk-lib';
 import {MetricsPublisherFunction} from './metrics-publisher-function';
-import {BitbucketJwtReceiverFunction} from './bitbucket-jwt-function';
 import {BitbucketMetricsRegisterFunction} from './bitbucket-metrics-register-function';
 
 export class BitbucketMetrics extends Construct {
