+++ handlers/src/bitbucket-metrics-register/bitbucket-auth-helper.ts
@@ -3,6 +3,9 @@ import {
   SecretsManagerClient,
 } from '@aws-sdk/client-secrets-manager';
 
+const AWS_REGION = process.env.AWS_REGION;
+const secretsManagerClient = new SecretsManagerClient({region: AWS_REGION});
+
 export interface Workspace {
   readonly name: string;
   readonly token: string;
@@ -13,18 +16,17 @@ export interface ScmData {
   readonly callBackUrl: string;
   readonly workspaces: Workspace[];
 }
-
-const currentRegion = process.env.AWS_REGION;
-const client = new SecretsManagerClient({region: currentRegion});
-export async function getScmData(secretName: string): Promise<ScmData | null> {
-  try {
-    const command = new GetSecretValueCommand({
-      SecretId: secretName,
-    });
-    const response = await client.send(command);
-    return JSON.parse(response!.SecretString!);
-  } catch (error) {
-    console.error('Error fetching Scm Data from secret. Exiting. ', error);
-    throw error;
+export class BitbucketAuthHelper {
+  public static async getScmData(secretName: string): Promise<ScmData | null> {
+    try {
+      const command = new GetSecretValueCommand({
+        SecretId: secretName,
+      });
+      const response = await secretsManagerClient.send(command);
+      return JSON.parse(response!.SecretString!);
+    } catch (error) {
+      console.error('Error fetching Scm Data from secret. Exiting. ', error);
+      throw error;
+    }
   }
 }
