+++ handlers/src/metrics-utilities/metrics-utilities.test.ts
@@ -1,22 +1,38 @@
-import {createRepositorySlug} from './metrics-utilities';
+import {MetricsUtilities} from './metrics-utilities';
 
-describe('createRepositorySlug', () => {
-  const testCases = [
-    {repositoryName: 'My Test Repository', expectedSlug: 'my-test-repository'},
-    {
-      repositoryName: 'My pecial Repoitory',
-      expectedSlug: 'my-pecial-repoitory',
-    },
-    {
-      repositoryName: 'curl_test',
-      expectedSlug: 'curl_test',
-    },
-  ];
+describe('metrics-utilities', () => {
+  describe('createRepositorySlug', () => {
+    const testCases = [
+      {
+        repositoryName: 'My Test Repository',
+        expectedSlug: 'my-test-repository',
+      },
+      {
+        repositoryName: 'My pecial Repoitory',
+        expectedSlug: 'my-pecial-repoitory',
+      },
+      {
+        repositoryName: 'curl_test',
+        expectedSlug: 'curl_test',
+      },
+    ];
 
-  testCases.forEach(({repositoryName, expectedSlug}) => {
-    it(`should correctly transform the repository name "${repositoryName}" into a slug "${expectedSlug}"`, () => {
-      const result = createRepositorySlug(repositoryName);
-      expect(result).toEqual(expectedSlug);
+    testCases.forEach(({repositoryName, expectedSlug}) => {
+      it(`should correctly transform the repository name "${repositoryName}" into a slug "${expectedSlug}"`, () => {
+        const result = MetricsUtilities.createRepositorySlug(repositoryName);
+        expect(result).toEqual(expectedSlug);
+      });
+    });
+  });
+
+  describe('readJsonToObject', () => {
+    it('should correctly read a JSON file and convert it to an object', () => {
+      const fileName = '../../test/data/test.json';
+      const expectedObject = {key: 'value'};
+
+      const result = MetricsUtilities.readJsonToObject(fileName);
+
+      expect(result).toEqual(expectedObject);
     });
   });
 });
