+++ handlers/src/metrics-utilities/metrics-utilities.ts
@@ -1,7 +1,22 @@
-export function createRepositorySlug(repositoryName: string): string {
-  // Convert to lowercase
-  const slug = repositoryName.toLowerCase();
+import path from 'path';
+import fs from 'fs';
 
-  // Replace spaces with hyphens
-  return slug.replace(/\s+/g, '-');
+export class MetricsUtilities {
+  public static createRepositorySlug(repositoryName: string): string {
+    // Convert to lowercase
+    const slug = repositoryName.toLowerCase();
+
+    // Replace spaces with hyphens
+    return slug.replace(/\s+/g, '-');
+  }
+
+  public static readJsonToObject(filename: string) {
+    const jsonFilePath = path.resolve(__dirname, filename);
+    try {
+      const data = fs.readFileSync(jsonFilePath, 'utf8');
+      return JSON.parse(data);
+    } catch (err) {
+      console.error(`Error reading file from disk: ${err}`);
+    }
+  }
 }
